requirements.txt
flask>=3.0
gunicorn>=22.0
biopython>=1.84
numpy>=1.26
pandas>=2.0
scipy>=1.11
scikit-bio>=0.6
primer3-py>=2.0
requests>=2.28
scikit-learn>=1.3
scikit-posthocs>=0.9
